Fix CurrencyCheckService and Test class
